rag_lab/requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
pydantic
python-multipart
streamlit
langchain
langchain-community
langchain-openai
langchain-chroma
chromadb
ragas
datasets
pandas
numpy
matplotlib
seaborn
koreanize-matplotlib
pypdf
